fix: upload chuck array

+++ src/shared/providers/StorageProvider/implementations/AmazonStorage/AmazonStorageProvider.ts
@@ -20,7 +20,7 @@ export class AmazonStorageProvider implements IStorageProvider {
     fileName,
     isPublic,
   }: FileStorage.Upload.Params): Promise<FileStorage.Upload.Result> {
-    // if (process.env.APP_HOST.includes('localhost')) return { url: 'edwq' };
+    if (process.env.APP_HOST.includes('localhost')) return { url: 'edwq' };
 
     const { Location: url } = await this.s3
       .upload({
@@ -69,7 +69,7 @@ export class AmazonStorageProvider implements IStorageProvider {
   async delete({
     fileName,
   }: FileStorage.Delete.Params): Promise<FileStorage.Delete.Result> {
-    // if (process.env.APP_HOST.includes('localhost')) return;
+    if (process.env.APP_HOST.includes('localhost')) return;
 
     await this.s3
       .deleteObject({ Bucket: this.bucket, Key: fileName })
